# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DMC API Documentation",
        "description": "Documentation de l'API pour l'application mobile DMC",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{productId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Liste des avis d un produit",
                "description": "Liste des avis d'un produit",
                "operationId": "28aa7c30dbbeadb77dc463f68ed1ff94",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}